Move SortOrderButton into components folder
